+++ packages/commonwealth/client/scripts/views/components/SublayoutHeader/DesktopHeader/DesktopHeader.tsx
@@ -16,7 +16,7 @@ import { CreateContentPopover } from 'views/menus/create_content_menu';
 import { HelpMenuPopover } from 'views/menus/help_menu';
 import { NotificationsMenuPopover } from 'views/menus/notifications_menu';
 
-import UserDropdown from '../UserDropdown';
+import UserDropdown from './UserDropdown';
 
 import './DesktopHeader.scss';
