+++ packages/commonwealth/client/scripts/views/components/SublayoutHeader/DesktopHeader/UserDropdown/UserDropdown.scss
@@ -1,4 +1,4 @@
-@import '../../../../../styles/shared';
+@import '../../../../../../styles/shared';
 
 .UserDropdown {
   width: 256px;
